osmApi: refactoring - move osmApiAuth
